refactor(node-drainer): type DrainScope and route the reconciler through the evaluator
Three review points from XRFXLP on #1714.

DrainScope is now a named string type rather than bare constants, since it is a small
fixed set used as a metric label.

The exported DrainScope function was dead: only its own test called it, because the
reconciler needed the entity as well and so re-derived both from PartialDrainEntity.
Replaced with DrainScopeFor, which returns the scope and the entity together, and the
reconciler now delegates to it. That removes the duplicate derivation and the dead code
in one change.

The partial-drain tests move into the existing evaluator_integration_test.go rather than
a separate file.
