+++ node-drainer/pkg/evaluator/evaluator_integration_test.go
@@ -20,6 +20,7 @@ import (
 	"testing"
 	"time"
 
+	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	v1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -662,4 +663,125 @@ func createHealthEventAnnotationsMap(healthEvents []*protos.HealthEvent) (string
 		return "", err
 	}
 	return string(annotationsMapBytes), nil
+}
+
+func healthEvent(action protos.RecommendedAction, entities ...*protos.Entity) *protos.HealthEvent {
+	return &protos.HealthEvent{
+		NodeName:          "node-1",
+		RecommendedAction: action,
+		EntitiesImpacted:  entities,
+	}
+}
+
+func gpuEntity(value string) *protos.Entity {
+	return &protos.Entity{EntityType: "GPU_UUID", EntityValue: value}
+}
+
+func TestDrainScopeFor_VaryingEventShapes_ReturnsEntityAndMatchingScope(t *testing.T) {
+	tests := []struct {
+		name       string
+		enabled    bool
+		event      *protos.HealthEvent
+		wantEntity string
+		wantScope  DrainScope
+	}{
+		{
+			name:      "partial drain disabled",
+			enabled:   false,
+			event:     healthEvent(protos.RecommendedAction_COMPONENT_RESET, gpuEntity("GPU-abc")),
+			wantScope: DrainScopeFull,
+		},
+		{
+			name:      "enabled but action is not COMPONENT_RESET",
+			enabled:   true,
+			event:     healthEvent(protos.RecommendedAction_RESTART_VM, gpuEntity("GPU-abc")),
+			wantScope: DrainScopeFull,
+		},
+		{
+			name:       "enabled with a usable GPU_UUID entity",
+			enabled:    true,
+			event:      healthEvent(protos.RecommendedAction_COMPONENT_RESET, gpuEntity("GPU-abc")),
+			wantEntity: "GPU-abc",
+			wantScope:  DrainScopePartial,
+		},
+		{
+			name:    "unsupported entity type is not usable",
+			enabled: true,
+			event: healthEvent(protos.RecommendedAction_COMPONENT_RESET,
+				&protos.Entity{EntityType: "NIC", EntityValue: "eth0"}),
+			wantScope: DrainScopeFull,
+		},
+		{
+			name:      "supported type with an empty value is not usable",
+			enabled:   true,
+			event:     healthEvent(protos.RecommendedAction_COMPONENT_RESET, gpuEntity("")),
+			wantScope: DrainScopeFull,
+		},
+		{
+			name:      "no entities at all",
+			enabled:   true,
+			event:     healthEvent(protos.RecommendedAction_COMPONENT_RESET),
+			wantScope: DrainScopeFull,
+		},
+		{
+			name:    "first usable entity wins when mixed",
+			enabled: true,
+			event: healthEvent(protos.RecommendedAction_COMPONENT_RESET,
+				&protos.Entity{EntityType: "NIC", EntityValue: "eth0"},
+				gpuEntity("GPU-def")),
+			wantEntity: "GPU-def",
+			wantScope:  DrainScopePartial,
+		},
+	}
+
+	for _, tc := range tests {
+		t.Run(tc.name, func(t *testing.T) {
+			entity := PartialDrainEntity(tc.event, tc.enabled)
+
+			if tc.wantEntity == "" {
+				assert.Nil(t, entity)
+			} else {
+				require.NotNil(t, entity)
+				assert.Equal(t, tc.wantEntity, entity.GetEntityValue())
+			}
+
+			gotScope, _ := DrainScopeFor(tc.event, tc.enabled)
+			assert.Equal(t, tc.wantScope, gotScope)
+		})
+	}
+}
+
+// The evaluator still treats "eligible but no usable entity" as an error, so a misconfigured
+// COMPONENT_RESET event fails loudly rather than silently draining the whole node.
+func TestShouldExecutePartialDrain_NoUsableEntity_ReturnsError(t *testing.T) {
+	e := &NodeDrainEvaluator{config: config.TomlConfig{PartialDrainEnabled: true}}
+
+	entity, err := e.shouldExecutePartialDrain(
+		healthEvent(protos.RecommendedAction_COMPONENT_RESET,
+			&protos.Entity{EntityType: "NIC", EntityValue: "eth0"}))
+
+	require.Error(t, err)
+	assert.Nil(t, entity)
+}
+
+func TestShouldExecutePartialDrain_UsableGPUEntity_ReturnsEntity(t *testing.T) {
+	e := &NodeDrainEvaluator{config: config.TomlConfig{PartialDrainEnabled: true}}
+
+	entity, err := e.shouldExecutePartialDrain(
+		healthEvent(protos.RecommendedAction_COMPONENT_RESET, gpuEntity("GPU-abc")))
+
+	require.NoError(t, err)
+	require.NotNil(t, entity)
+	assert.Equal(t, "GPU-abc", entity.GetEntityValue())
+}
+
+// A non-candidate event returns no entity and no error, so full drain proceeds normally.
+func TestShouldExecutePartialDrain_PartialDrainDisabled_ReturnsNilWithoutError(t *testing.T) {
+	e := &NodeDrainEvaluator{config: config.TomlConfig{PartialDrainEnabled: false}}
+
+	entity, err := e.shouldExecutePartialDrain(
+		healthEvent(protos.RecommendedAction_COMPONENT_RESET, gpuEntity("GPU-abc")))
+
+	require.NoError(t, err)
+	assert.Nil(t, entity)
 }
